+++ source/Processing.Infrastructure/Configuration/EventPublishingRegistration.cs
@@ -17,6 +17,7 @@ using Azure.Messaging.ServiceBus;
 using Contracts.IntegrationEvents;
 using Processing.Application.Common;
 using Processing.Infrastructure.Configuration.EventPublishing;
+using Processing.Infrastructure.Configuration.EventPublishing.AzureServiceBus;
 using SimpleInjector;
 
 namespace Processing.Infrastructure.Configuration
